chore: 同步 DefaultScatter

+++ utils-support-deeplearning-parent/utils-support-deeplearning-onnx-starter/src/main/java/com/chua/deeplearning/support/onnx/yoloworld/YoloWorldDetectorTranslator.java
@@ -268,7 +268,7 @@ public class YoloWorldDetectorTranslator implements Translator<Image, DetectedOb
         }
 
         // NMS
-        List<Integer> keep = nms(boxes, probabilities, nmsThreshold);
+        List<Integer> keep = nms(boxes, probabilities, (float) nmsThreshold);
 
         List<String> finalNames = new ArrayList<>();
         List<Double> finalProbs = new ArrayList<>();
